#767: cleaned up old parser
