+++ packages/analysis/package.json
@@ -5,8 +5,8 @@
   "author": "Masking Technology <info@masking.tech> (https://jitar.dev)",
   "license": "MIT",
   "type": "module",
-  "types": "dist/indexNew.d.ts",
-  "exports": "./dist/indexNew.js",
+  "types": "dist/index.d.ts",
+  "exports": "./dist/index.js",
   "private": true,
   "scripts": {
     "test": "vitest run",
