+++ packages/analysis/src/dynamic/index.ts
@@ -1,3 +1,2 @@
 
 export { default as Reflector } from './Reflector';
-export { default as ReflectorNew } from './ReflectorNew';
