+++ packages/analysis/src/index.ts
@@ -1,9 +1,11 @@
 
 export {
-    ESAlias, ESArray, ESClass, ESDeclaration, ESDestructuredArray, ESDestructuredObject, ESDestructuredValue,
-    ESExport, ESExpression, ESField, ESFunction, ESGenerator, ESGetter, ESImport, ESMember, ESModule,
-    ESObject, ESParameter, ESSetter
-} from './models';
+    ESBinding, ESIdentifierBinding, ESArrayBinding, ESObjectBinding, ESBindingElement,
+    ESDeclaration, ESBlock, ESExpression, ESVariable,
+    ESFunction, ESArrowFunction, ESGeneratorFunction, ESParameter,
+    ESClass, ESClassMember, ESField, ESMethod, ESGeneratorMethod, ESConstructor, ESGetter, ESSetter,
+    ESModule, ESModuleMember, ESExport, ESImport, ESStatement
+} from './model';
 
 export { Reflector } from './dynamic';
 export { Parser } from './static';
