--- packages/analysis/src/indexNew.ts
@@ -1,11 +1,0 @@
-
-export {
-    ESBinding, ESIdentifierBinding, ESArrayBinding, ESObjectBinding, ESBindingElement,
-    ESDeclaration, ESBlock, ESExpression, ESVariable,
-    ESFunction, ESArrowFunction, ESGeneratorFunction, ESParameter,
-    ESClass, ESClassMember, ESField, ESMethod, ESGeneratorMethod, ESConstructor, ESGetter, ESSetter,
-    ESModule, ESModuleMember, ESExport, ESImport, ESStatement
-} from './model';
-
-export { ReflectorNew as Reflector } from './dynamic';
-export { ParserNew as Parser } from './static';
