--- packages/analysis/src/models/ESArray.ts
@@ -1,7 +1,0 @@
-
-import ESValue from './ESValue.js';
-
-export default class ESArray extends ESValue
-{
-    
-}
