--- packages/analysis/src/models/ESDestructuredArray.ts
@@ -1,10 +1,0 @@
-
-import ESDestructuredValue from './ESDestructuredValue.js';
-
-export default class ESDestructuredArray extends ESDestructuredValue
-{
-    toString(): string
-    {
-        return `[ ${super.toString()} ]`;
-    }
-}
