--- packages/analysis/src/models/ESDestructuredObject.ts
@@ -1,10 +1,0 @@
-
-import ESDestructuredValue from './ESDestructuredValue.js';
-
-export default class ESDestructuredObject extends ESDestructuredValue
-{
-    toString(): string
-    {
-        return `{ ${super.toString()} }`;
-    }
-}
