--- packages/analysis/src/models/ESExpression.ts
@@ -1,7 +1,0 @@
-
-import ESValue from './ESValue.js';
-
-export default class ESExpression extends ESValue
-{
-    
-}
