--- packages/analysis/src/models/ESGetter.ts
@@ -1,10 +1,0 @@
-
-import ESFunction from './ESFunction.js';
-
-export default class ESGetter extends ESFunction
-{
-    toString(): string
-    {
-        return `get ${super.toString()}`;
-    }
-}
