--- packages/analysis/src/models/ESIdentifier.ts
@@ -1,7 +1,0 @@
-
-import ESDestructuredArray from './ESDestructuredArray.js';
-import ESDestructuredObject from './ESDestructuredObject.js';
-
-type ESIdentifier = string | ESDestructuredArray | ESDestructuredObject
-
-export default ESIdentifier;
