--- packages/analysis/src/models/ESObject.ts
@@ -1,7 +1,0 @@
-
-import ESValue from './ESValue.js';
-
-export default class ESObject extends ESValue
-{
-    
-}
