--- packages/analysis/src/models/ESParameter.ts
@@ -1,8 +1,0 @@
-
-import ESDestructuredArray from './ESDestructuredArray.js';
-import ESField from './ESField.js';
-import ESDestructuredObject from './ESDestructuredObject.js';
-
-type ESParameter = ESField | ESDestructuredObject | ESDestructuredArray;
-
-export default ESParameter;
