--- packages/analysis/src/models/ESSetter.ts
@@ -1,10 +1,0 @@
-
-import ESFunction from './ESFunction.js';
-
-export default class ESSetter extends ESFunction
-{
-    toString(): string
-    {
-        return `set ${super.toString()}`;
-    }
-}
