--- packages/analysis/src/models/ESValue.ts
@@ -1,17 +1,0 @@
-
-export default class ESValue
-{
-    readonly #definition: string;
-
-    constructor(definition: string)
-    {
-        this.#definition = definition;
-    }
-
-    get definition() { return this.#definition; }
-
-    toString(): string
-    {
-        return this.#definition;
-    }
-}
