--- packages/analysis/src/models/index.ts
@@ -1,23 +1,0 @@
-
-export { default as ESAlias } from './ESAlias';
-export { default as ESArray } from './ESArray';
-export { default as ESClass } from './ESClass';
-export { default as ESDeclaration } from './ESDeclaration';
-export { default as ESDestructuredArray } from './ESDestructuredArray';
-export { default as ESDestructuredObject } from './ESDestructuredObject';
-export { default as ESDestructuredValue } from './ESDestructuredValue';
-export { default as ESExport } from './ESExport';
-export { default as ESExpression } from './ESExpression';
-export { default as ESField } from './ESField';
-export { default as ESFunction } from './ESFunction';
-export { default as ESGenerator } from './ESGenerator';
-export { default as ESGetter } from './ESGetter';
-export { default as ESIdentifier } from './ESIdentifier';
-export { default as ESImport } from './ESImport';
-export { default as ESMember } from './ESMember';
-export { default as ESModule } from './ESModule';
-export { default as ESObject } from './ESObject';
-export { default as ESParameter } from './ESParameter';
-export { default as ESScope } from './ESScope';
-export { default as ESSetter } from './ESSetter';
-export { default as ESValue } from './ESValue';
