+++ packages/analysis/src/static/index.ts
@@ -12,6 +12,5 @@ export * from './definitions/Scope';
 export * from './definitions/TokenType';
 export * from './definitions/Whitespace';
 
-export { default as Lexer } from './Lexer';
+export { default as Lexer} from './Lexer';
 export { default as Parser } from './Parser';
-export { default as ParserNew } from './ParserNew';
