--- packages/analysis/test/models/fixtures/index.ts
@@ -1,4 +1,0 @@
-
-export * from './esClass.fixture';
-export * from './esModule.fixture';
-export * from './esScope.fixture';
