+++ packages/analysis/test/static/Parser.spec.ts
@@ -1,57 +1,63 @@
 
 import { describe, expect, it } from 'vitest';
 
-import { ESArray, ESObject, ESExpression, ESField, ESGenerator, ESDestructuredArray, ESDestructuredObject } from '../../src/models';
+import {
+    ESIdentifierBinding, ESArrayBinding, ESObjectBinding, ESBindingElement,
+    ESExpression,
+    ESFunction, ESArrowFunction, ESGeneratorFunction,
+    ESClass, ESGeneratorMethod, ESConstructor
+} from '../../src/model';
+
 import { Parser } from '../../src/static';
 
-import { VALUES, IMPORTS, EXPORTS, DECLARATIONS, FUNCTIONS, CLASSES, MODULES } from './fixtures';
+import { VALUES, IMPORTS, EXPORTS, DECLARATIONS, FUNCTIONS, CLASSES, MODULES, MODULES_STRINGS } from './fixtures';
 
 const parser = new Parser();
 
 describe('parser/Parser', () =>
 {
-    describe('.parseValue(code)', () =>
+    describe('.parseStatement(code)', () =>
     {
         it('should parse an array', () =>
         {
-            const value = parser.parseValue(VALUES.ARRAY);
-            expect(value).toBeInstanceOf(ESArray);
-            expect(value.definition).toBe('[ 1 , "foo" , false , new Person ( "Peter" , 42 ) , { a : 1 , b : 2 } ]');
+            const expression = parser.parseStatement(VALUES.ARRAY);
+            expect(expression).toBeInstanceOf(ESExpression);
+            expect(expression.toString(false)).toEqual('[ 1 , "foo" , false , new Person ( "Peter" , 42 ) , { a : 1 , b : 2 } ]');
         });
 
         it('should parse an object', () =>
         {
-            const value = parser.parseValue(VALUES.OBJECT);
-            expect(value).toBeInstanceOf(ESObject);
-            expect(value.definition).toBe('{ key1 : "value1" , "key2" : new Person ( ) .toString ( ) }');
+            const expression = parser.parseStatement(VALUES.OBJECT);
+            expect(expression).toBeInstanceOf(ESExpression);
+            expect(expression.toString(false)).toEqual('{ key1 : "value1" , "key2" : new Person ( ) .toString ( ) }');
         });
 
         it('should parse an expression', () =>
         {
-            const value = parser.parseValue(VALUES.EXPRESSION);
-            expect(value).toBeInstanceOf(ESExpression);
-            expect(value.definition).toBe('new Number ( Math.ceil ( Math.random ( ) ) + 10 ) .toString ( )');
+            const expression = parser.parseStatement(VALUES.EXPRESSION);
+            expect(expression).toBeInstanceOf(ESExpression);
+            expect(expression.toString(false)).toEqual('new Number ( Math.ceil ( Math.random ( ) ) + 10 ) .toString ( )');
         });
 
         it('should parse a grouped expression', () =>
         {
-            const value = parser.parseValue(VALUES.EXPRESSION_GROUP);
-            expect(value).toBeInstanceOf(ESExpression);
-            expect(value.definition).toBe('( a + b ) * c');
+            const expression = parser.parseStatement(VALUES.EXPRESSION_GROUP);
+            expect(expression).toBeInstanceOf(ESExpression);
+            expect(expression.toString(false)).toEqual('( a + b ) * c');
         });
 
         it('should parse an if...else expression', () =>
         {
-            const value = parser.parseValue(VALUES.IF_ELSE);
-            expect(value).toBeInstanceOf(ESExpression);
-            expect(value.definition).toBe('if ( true ) { return "value1" ; } else { return "value2" ; }');
+            const expression = parser.parseStatement(VALUES.IF_ELSE);
+            expect(expression).toBeInstanceOf(ESExpression);
+            expect(expression.toString(false)).toEqual('if ( true ) { return "value1" ; } else { return "value2" ; }');
         });
 
         it('should parse an try...catch...finally expression', () =>
         {
-            const value = parser.parseValue(VALUES.TRY_CATCH_FINALLY);
-            expect(value).toBeInstanceOf(ESExpression);
-            expect(value.definition).toBe('try { sum ( 1 , 2 ) ; } catch ( error ) { console.error ( error ) ; } finally { console.log ( "finally" ) ; }');
+            const expression = parser.parseStatement(VALUES.TRY_CATCH_FINALLY);
+            expect(expression).toBeInstanceOf(ESExpression);
+            expect(expression.toString(false)).toEqual('try { sum ( 1 , 2 ) ; } catch ( error ) { console.error ( error ) ; } finally { console.log ( "finally" ) ; }');
         });
     });
 
@@ -60,104 +66,104 @@ describe('parser/Parser', () =>
         it('should parse loading a module', () =>
         {
             const imported = parser.parseImport(IMPORTS.LOAD);
-            expect(imported.members.length).toBe(0);
-            expect(imported.from).toBe("'module'");
+            expect(imported.members).toHaveLength(0);
+            expect(imported.from).toEqual('module');
         });
 
         it('should parse importing a single member', () =>
         {
             const imported = parser.parseImport(IMPORTS.IMPORT);
-            expect(imported.members.length).toBe(1);
-            expect(imported.from).toBe("'module'");
+            expect(imported.members).toHaveLength(1);
+            expect(imported.from).toEqual('module');
 
             const member = imported.members[0];
-            expect(member.name).toBe('member');
-            expect(member.as).toBe('member');
+            expect(member.identifier).toEqual('member');
+            expect(member.alias).toBeUndefined();
         });
 
         it('should parse importing a single member with alias', () =>
         {
             const imported = parser.parseImport(IMPORTS.IMPORT_AS);
-            expect(imported.members.length).toBe(1);
-            expect(imported.from).toBe("'module'");
+            expect(imported.members).toHaveLength(1);
+            expect(imported.from).toEqual('module');
 
             const member = imported.members[0];
-            expect(member.name).toBe('member');
-            expect(member.as).toBe('alias');
+            expect(member.identifier).toEqual('member');
+            expect(member.alias).toEqual('alias');
         });
 
         it('should parse importing all members', () =>
         {
             const imported = parser.parseImport(IMPORTS.IMPORT_ALL);
-            expect(imported.members.length).toBe(1);
-            expect(imported.from).toBe("'module'");
+            expect(imported.members).toHaveLength(1);
+            expect(imported.from).toEqual('module');
 
             const member = imported.members[0];
-            expect(member.name).toBe('*');
-            expect(member.as).toBe('name');
+            expect(member.identifier).toEqual('*');
+            expect(member.alias).toEqual('name');
         });
 
         it('should parse importing the default member', () =>
         {
             const imported = parser.parseImport(IMPORTS.IMPORT_DEFAULT);
-            expect(imported.members.length).toBe(1);
-            expect(imported.from).toBe("'module'");
+            expect(imported.members).toHaveLength(1);
+            expect(imported.from).toEqual('module');
 
             const member = imported.members[0];
-            expect(member.name).toBe('default');
-            expect(member.as).toBe('name');
+            expect(member.identifier).toEqual('default');
+            expect(member.alias).toEqual('name');
         });
 
         it('should parse importing the default with another member', () =>
         {
             const imported = parser.parseImport(IMPORTS.IMPORT_DEFAULT_MEMBER);
-            expect(imported.members.length).toBe(2);
-            expect(imported.from).toBe("'module'");
+            expect(imported.members).toHaveLength(2);
+            expect(imported.from).toEqual('module');
 
             const first = imported.members[0];
-            expect(first.name).toBe('default');
-            expect(first.as).toBe('name');
+            expect(first.identifier).toEqual('default');
+            expect(first.alias).toEqual('name');
 
             const second = imported.members[1];
-            expect(second.name).toBe('member');
-            expect(second.as).toBe('member');
+            expect(second.identifier).toEqual('member');
+            expect(second.alias).toBeUndefined();
         });
 
         it('should parse importing the default with another member with alias', () =>
         {
             const imported = parser.parseImport(IMPORTS.IMPORT_DEFAULT_MEMBER_AS);
-            expect(imported.members.length).toBe(2);
-            expect(imported.from).toBe("'module'");
+            expect(imported.members).toHaveLength(2);
+            expect(imported.from).toEqual('module');
 
             const first = imported.members[0];
-            expect(first.name).toBe('default');
-            expect(first.as).toBe('name');
+            expect(first.identifier).toEqual('default');
+            expect(first.alias).toEqual('name');
 
             const second = imported.members[1];
-            expect(second.name).toBe('member');
-            expect(second.as).toBe('alias');
+            expect(second.identifier).toEqual('member');
+            expect(second.alias).toEqual('alias');
         });
 
         it('should parse importing the default with other members and aliases', () =>
         {
             const imported = parser.parseImport(IMPORTS.IMPORT_MULTIPLE_MEMBERS_AS);
-            expect(imported.members.length).toBe(2);
-            expect(imported.from).toBe("'module'");
+            expect(imported.members).toHaveLength(2);
+            expect(imported.from).toEqual('module');
 
             const first = imported.members[0];
-            expect(first.name).toBe('name');
-            expect(first.as).toBe('name');
+            expect(first.identifier).toEqual('name');
+            expect(first.alias).toBeUndefined();
 
             const second = imported.members[1];
-            expect(second.name).toBe('member');
-            expect(second.as).toBe('alias');
+            expect(second.identifier).toEqual('member');
+            expect(second.alias).toEqual('alias');
         });
 
         it('should parse importing modules dynamically', () =>
         {
             const imported = parser.parseImport(IMPORTS.IMPORT_DYNAMIC);
-            expect(imported.members.length).toBe(0);
-            expect(imported.from).toBe("'module'");
+            expect(imported.members).toHaveLength(0);
+            expect(imported.from).toEqual('module');
         });
     });
 
@@ -166,265 +172,282 @@ describe('parser/Parser', () =>
         it('should parse exporting a single member', () =>
         {
             const exported = parser.parseExport(EXPORTS.EXPORT);
-            expect(exported.members.length).toBe(1);
-            expect(exported.from).toBe(undefined);
+            expect(exported.members).toHaveLength(1);
+            expect(exported.from).toBeUndefined();
 
             const member = exported.members[0];
-            expect(member.name).toBe('member');
-            expect(member.as).toBe('member');
+            expect(member.identifier).toEqual('member');
+            expect(member.alias).toBeUndefined();
         });
 
         it('should parse exporting a single member with alias', () =>
         {
             const exported = parser.parseExport(EXPORTS.EXPORT_AS);
-            expect(exported.members.length).toBe(1);
-            expect(exported.from).toBe(undefined);
+            expect(exported.members).toHaveLength(1);
+            expect(exported.from).toBeUndefined();
 
             const member = exported.members[0];
-            expect(member.name).toBe('member');
-            expect(member.as).toBe('alias');
+            expect(member.identifier).toEqual('member');
+            expect(member.alias).toEqual('alias');
         });
 
         it('should parse exporting a default', () =>
         {
             const exported = parser.parseExport(EXPORTS.EXPORT_DEFAULT);
-            expect(exported.members.length).toBe(1);
-            expect(exported.from).toBe(undefined);
+            expect(exported.members).toHaveLength(1);
+            expect(exported.from).toBeUndefined();
 
             const member = exported.members[0];
-            expect(member.name).toBe('name');
-            expect(member.as).toBe('default');
+            expect(member.identifier).toEqual('name');
+            expect(member.alias).toEqual('default');
         });
 
         it('should parse exporting multiple members', () =>
         {
             const exported = parser.parseExport(EXPORTS.EXPORT_MULTIPLE);
-            expect(exported.members.length).toBe(2);
-            expect(exported.from).toBe(undefined);
+            expect(exported.members).toHaveLength(2);
+            expect(exported.from).toBeUndefined();
 
             const first = exported.members[0];
-            expect(first.name).toBe('name');
-            expect(first.as).toBe('name');
+            expect(first.identifier).toEqual('name');
+            expect(first.alias).toBeUndefined();
 
             const second = exported.members[1];
-            expect(second.name).toBe('member');
-            expect(second.as).toBe('member');
+            expect(second.identifier).toEqual('member');
+            expect(second.alias).toBeUndefined();
         });
 
         it('should parse exporting multiple members with alias', () =>
         {
             const exported = parser.parseExport(EXPORTS.EXPORT_MULTIPLE_AS);
-            expect(exported.members.length).toBe(2);
-            expect(exported.from).toBe(undefined);
+            expect(exported.members).toHaveLength(2);
+            expect(exported.from).toBeUndefined();
 
             const first = exported.members[0];
-            expect(first.name).toBe('name');
-            expect(first.as).toBe('name');
+            expect(first.identifier).toEqual('name');
+            expect(first.alias).toBeUndefined();
 
             const second = exported.members[1];
-            expect(second.name).toBe('member');
-            expect(second.as).toBe('alias');
+            expect(second.identifier).toEqual('member');
+            expect(second.alias).toEqual('alias');
         });
 
         it('should parse exporting a class declaration', () =>
         {
             const exported = parser.parseExport(EXPORTS.EXPORT_CLASS_DECLARATION);
-            expect(exported.members.length).toBe(1);
-            expect(exported.from).toBe(undefined);
+            expect(exported.members).toHaveLength(1);
+            expect(exported.from).toBeUndefined();
 
             const member = exported.members[0];
-            expect(member.name).toBe('name');
-            expect(member.as).toBe('name');
+            expect(member.identifier).toEqual('name');
+            expect(member.alias).toBeUndefined();
         });
 
         it('should parse exporting a function declaration', () =>
         {
             const exported = parser.parseExport(EXPORTS.EXPORT_FUNCTION_DECLARATION);
-            expect(exported.members.length).toBe(1);
-            expect(exported.from).toBe(undefined);
+            expect(exported.members).toHaveLength(1);
+            expect(exported.from).toBeUndefined();
 
             const member = exported.members[0];
-            expect(member.name).toBe('name');
-            expect(member.as).toBe('name');
+            expect(member.identifier).toEqual('name');
+            expect(member.alias).toBeUndefined();
         });
 
         it('should parse exporting a async function declaration', () =>
         {
             const exported = parser.parseExport(EXPORTS.EXPORT_FIELD_DECLARATION);
-            expect(exported.members.length).toBe(1);
-            expect(exported.from).toBe(undefined);
+            expect(exported.members).toHaveLength(1);
+            expect(exported.from).toBeUndefined();
 
             const member = exported.members[0];
-            expect(member.name).toBe('name');
-            expect(member.as).toBe('name');
+            expect(member.identifier).toEqual('name');
+            expect(member.alias).toBeUndefined();
         });
 
         it('should parse reexporting a module', () =>
         {
             const exported = parser.parseExport(EXPORTS.REEXPORT_ALL);
-            expect(exported.members.length).toBe(1);
-            expect(exported.from).toBe("'module'");
+            expect(exported.members).toHaveLength(1);
+            expect(exported.from).toEqual('module');
 
             const member = exported.members[0];
-            expect(member.name).toBe('');
-            expect(member.as).toBe('');
+            expect(member.identifier).toEqual('');
+            expect(member.alias).toBeUndefined();
         });
 
         it('should parse reexporting a member', () =>
         {
             const exported = parser.parseExport(EXPORTS.REEXPORT_MEMBER);
-            expect(exported.members.length).toBe(1);
-            expect(exported.from).toBe("'module'");
+            expect(exported.members).toHaveLength(1);
+            expect(exported.from).toEqual('module');
 
             const member = exported.members[0];
-            expect(member.name).toBe('member');
-            expect(member.as).toBe('member');
+            expect(member.identifier).toEqual('member');
+            expect(member.alias).toBeUndefined();
         });
     });
 
-    describe('.parseDeclaration(code)', () =>
+    describe('.parseVariable(code)', () =>
     {
         it('should parse an empty declaration', () =>
         {
-            const declaration = parser.parseDeclaration(DECLARATIONS.EMPTY);
-
-            expect(declaration.name).toBe('name');
-            expect(declaration.value).toBe(undefined);
+            const variable = parser.parseVariable(DECLARATIONS.EMPTY);
+            expect(variable.identifier).toEqual('name');
+            expect(variable.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(variable.type).toEqual('let');
+            expect(variable.initializer).toBeUndefined();
         });
 
-        it('should parse a const declaration', () =>
+        it('should parse a const variable', () =>
         {
-            const declaration = parser.parseDeclaration(DECLARATIONS.CONST);
-
-            expect(declaration.name).toBe('name');
-            expect(declaration.value).toBeInstanceOf(ESExpression);
-            expect(declaration.value?.definition).toBe("'const'");
+            const variable = parser.parseVariable(DECLARATIONS.CONST);
+            expect(variable.identifier).toEqual('name');
+            expect(variable.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(variable.type).toEqual('const');
+            expect(variable.initializer).toBeInstanceOf(ESExpression);
+            expect(variable.initializer?.toString(false)).toEqual("'const'");
         });
 
         it('should parse a let declaration with value', () =>
         {
-            const declaration = parser.parseDeclaration(DECLARATIONS.LET);
-
-            expect(declaration.name).toBe('name');
-            expect(declaration.value).toBeInstanceOf(ESExpression);
-            expect(declaration.value?.definition).toBe("'let'");
+            const variable = parser.parseVariable(DECLARATIONS.LET);
+            expect(variable.identifier).toEqual('name');
+            expect(variable.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(variable.type).toEqual('let');
+            expect(variable.initializer).toBeInstanceOf(ESExpression);
+            expect(variable.initializer?.toString(false)).toEqual("'let'");
         });
 
         it('should parse a var declaration with value', () =>
         {
-            const declaration = parser.parseDeclaration(DECLARATIONS.VAR);
-
-            expect(declaration.name).toBe('name');
-            expect(declaration.value).toBeInstanceOf(ESExpression);
-            expect(declaration.value?.definition).toBe("'var'");
+            const variable = parser.parseVariable(DECLARATIONS.VAR);
+            expect(variable.identifier).toEqual('name');
+            expect(variable.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(variable.type).toEqual('var');
+            expect(variable.initializer).toBeInstanceOf(ESExpression);
+            expect(variable.initializer?.toString(false)).toEqual("'var'");
         });
 
         it('should parse a declaration with multiple declarations', () =>
         {
-            const declaration = parser.parseDeclaration(DECLARATIONS.MULTIPLE);
-
-            expect(declaration.name).toBe('name1');
-            expect(declaration.value).toBeInstanceOf(ESExpression);
-            expect(declaration.value?.definition).toBe('( 1 + 2 ) * 3');
+            const variable = parser.parseVariable(DECLARATIONS.MULTIPLE);
+            expect(variable.identifier).toEqual('name1');
+            expect(variable.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(variable.type).toEqual('let');
+            expect(variable.initializer).toBeInstanceOf(ESExpression);
+            expect(variable.initializer?.toString(false)).toEqual('( 1 + 2 ) * 3');
         });
 
         it('should parse a declaration with an expression', () =>
         {
-            const declaration = parser.parseDeclaration(DECLARATIONS.EXPRESSION);
-
-            expect(declaration.name).toBe('number');
-            expect(declaration.value).toBeInstanceOf(ESExpression);
-            expect(declaration.value?.definition).toBe("new Number ( Math.ceil ( Math.random ( ) ) + 10 ) .toString ( )");
+            const variable = parser.parseVariable(DECLARATIONS.EXPRESSION);
+            expect(variable.identifier).toEqual('number');
+            expect(variable.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(variable.type).toEqual('const');
+            expect(variable.initializer).toBeInstanceOf(ESExpression);
+            expect(variable.initializer?.toString(false)).toEqual('new Number ( Math.ceil ( Math.random ( ) ) + 10 ) .toString ( )');
         });
 
         it('should parse a declaration with an array value', () =>
         {
-            const declaration = parser.parseDeclaration(DECLARATIONS.ARRAY);
-
-            expect(declaration.name).toBe('array');
-            expect(declaration.value).toBeInstanceOf(ESArray);
-            expect(declaration.value?.definition).toBe("[ 'value1' , 'value2' ]");
+            const variable = parser.parseVariable(DECLARATIONS.ARRAY);
+            expect(variable.identifier).toEqual('array');
+            expect(variable.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(variable.type).toEqual('const');
+            expect(variable.initializer).toBeInstanceOf(ESExpression);
+            expect(variable.initializer?.toString(false)).toEqual("[ 'value1' , 'value2' ]");
         });
 
         it('should parse a declaration with an array value', () =>
         {
-            const declaration = parser.parseDeclaration(DECLARATIONS.OBJECT);
-
-            expect(declaration.name).toBe('object');
-            expect(declaration.value).toBeInstanceOf(ESObject);
-            expect(declaration.value?.definition).toBe("{ key1 : 'value1' , key2 : 'value2' }");
+            const variable = parser.parseVariable(DECLARATIONS.OBJECT);
+            expect(variable.identifier).toEqual('object');
+            expect(variable.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(variable.type).toEqual('const');
+            expect(variable.initializer).toBeInstanceOf(ESExpression);
+            expect(variable.initializer?.toString(false)).toEqual("{ key1 : 'value1' , key2 : 'value2' }");
         });
 
         it('should parse a declaration with a regex value', () =>
         {
-            const declaration = parser.parseDeclaration(DECLARATIONS.REGEX);
-
-            expect(declaration.name).toBe('regex');
-            expect(declaration.value).toBeInstanceOf(ESExpression);
-            expect(declaration.value?.definition).toBe("/regex/g");
+            const variable = parser.parseVariable(DECLARATIONS.REGEX);
+            expect(variable.identifier).toEqual('regex');
+            expect(variable.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(variable.type).toEqual('const');
+            expect(variable.initializer).toBeInstanceOf(ESExpression);
+            expect(variable.initializer?.toString(false)).toEqual("/regex/g");
         });
 
         it('should parse a declaration that is destructuring an array', () =>
         {
-            const declaration = parser.parseDeclaration(DECLARATIONS.DESTRUCTURING_ARRAY);
+            const variable = parser.parseVariable(DECLARATIONS.DESTRUCTURING_ARRAY);
+            expect(variable.identifier).toEqual('[ value1, value2 = true ]');
+            expect(variable.binding).toBeInstanceOf(ESArrayBinding);
+            expect(variable.type).toEqual('const');
+            expect(variable.initializer).toBeInstanceOf(ESExpression);
+            expect(variable.initializer?.toString(false)).toEqual('array');
 
-            expect(declaration.name).toBe('[ value1 , value2 = true ]');
-            expect(declaration.value).toBeInstanceOf(ESExpression);
-            expect(declaration.identifier).toBeInstanceOf(ESDestructuredArray);
-
-            const identifier = declaration.identifier as ESDestructuredArray;
-            expect(identifier.members.length).toBe(2);
+            const binding = variable.binding as ESArrayBinding;
+            expect(binding.elements).toHaveLength(2);
             
-            const firstMember = identifier.members[0] as ESField;
-            expect(firstMember).toBeInstanceOf(ESField);
-            expect(firstMember.name).toBe('value1');
-            expect(firstMember.value).toBe(undefined);
+            const firstElement = binding.elements[0] as ESBindingElement;
+            expect(firstElement).toBeInstanceOf(ESBindingElement);
+            expect(firstElement.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(firstElement.binding.toString()).toEqual('value1');
+            expect(firstElement.initializer).toBeUndefined();
 
-            const secondMember = identifier.members[1] as ESField;
-            expect(secondMember).toBeInstanceOf(ESField);
-            expect(secondMember.name).toBe('value2');
-            expect(secondMember.value).toBeInstanceOf(ESExpression);
-            expect(secondMember.value?.definition).toBe('true');
+            const secondElement = binding.elements[1] as ESBindingElement;
+            expect(secondElement).toBeInstanceOf(ESBindingElement);
+            expect(secondElement.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(secondElement.binding.toString()).toEqual('value2');
+            expect(secondElement.initializer).toBeInstanceOf(ESExpression);
+            expect(secondElement.initializer?.toString(false)).toEqual('true');
         });
 
         it('should parse a declaration that is destructuring an object', () =>
         {
-            const declaration = parser.parseDeclaration(DECLARATIONS.DESTRUCTURING_OBJECT);
+            const variable = parser.parseVariable(DECLARATIONS.DESTRUCTURING_OBJECT);
+            expect(variable.identifier).toEqual('{ key1, key2 = false }');
+            expect(variable.binding).toBeInstanceOf(ESObjectBinding);
+            expect(variable.type).toEqual('const');
+            expect(variable.initializer).toBeInstanceOf(ESExpression);
+            expect(variable.initializer?.toString(false)).toEqual('object');
 
-            expect(declaration.name).toBe('{ key1 , key2 = false }');
-            expect(declaration.value).toBeInstanceOf(ESExpression);
-            expect(declaration.identifier).toBeInstanceOf(ESDestructuredObject);
+            const binding = variable.binding as ESObjectBinding;
+            expect(binding.elements).toHaveLength(2);
 
-            const identifier = declaration.identifier as ESDestructuredArray;
-            expect(identifier.members.length).toBe(2);
+            const firstElement = binding.elements[0] as ESBindingElement;
+            expect(firstElement.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(firstElement.binding.toString()).toEqual('key1');
+            expect(firstElement.initializer).toBeUndefined();
 
-            const firstMember = identifier.members[0] as ESField;
-            expect(firstMember).toBeInstanceOf(ESField);
-            expect(firstMember.name).toBe('key1');
-
-            const secondMember = identifier.members[1] as ESField;
-            expect(secondMember).toBeInstanceOf(ESField);
-            expect(secondMember.name).toBe('key2');
-            expect(secondMember.value).toBeInstanceOf(ESExpression);
-            expect(secondMember.value?.definition).toBe('false');
+            const secondElement = binding.elements[1] as ESBindingElement;
+            expect(secondElement).toBeInstanceOf(ESBindingElement);
+            expect(secondElement.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(secondElement.binding.toString()).toEqual('key2');
+            expect(secondElement.initializer).toBeInstanceOf(ESExpression);
+            expect(secondElement.initializer?.toString(false)).toEqual('false');
         });
 
         it('should parse a declaration with a non reserved keyword as name', () =>
         {
-            const declaration = parser.parseDeclaration(DECLARATIONS.KEYWORD_AS_NAME);
-
-            expect(declaration.name).toBe('as');
-            expect(declaration.value).toBeInstanceOf(ESExpression);
-            expect(declaration.value?.definition).toBe("'value'");
+            const variable = parser.parseVariable(DECLARATIONS.KEYWORD_AS_NAME);
+            expect(variable.identifier).toEqual('as');
+            expect(variable.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(variable.type).toEqual('const');
+            expect(variable.initializer).toBeInstanceOf(ESExpression);
+            expect(variable.initializer?.toString(false)).toEqual("'value'");
         });
 
         it('should parse a declaration that refers to a non reserved keyword as value', () =>
         {
-            const declaration = parser.parseDeclaration(DECLARATIONS.KEYWORD_AS_VALUE);
-
-            expect(declaration.name).toBe('alias');
-            expect(declaration.value).toBeInstanceOf(ESExpression);
-            expect(declaration.value?.definition).toBe('as');
+            const variable = parser.parseVariable(DECLARATIONS.KEYWORD_AS_VALUE);
+            expect(variable.identifier).toEqual('alias');
+            expect(variable.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(variable.type).toEqual('const');
+            expect(variable.initializer).toBeInstanceOf(ESExpression);
+            expect(variable.initializer?.toString(false)).toEqual('as');
         });
     });
 
@@ -433,333 +456,396 @@ describe('parser/Parser', () =>
         it('should parse a simple function declaration', () =>
         {
             const funktion = parser.parseFunction(FUNCTIONS.DECLARATION);
-
-            expect(funktion.name).toBe('name');
-            expect(funktion.isAsync).toBe(false);
-            expect(funktion.parameters.length).toBe(0);
-            expect(funktion.body).toBe('{ }');
+            expect(funktion.identifier).toEqual('name');
+            expect(funktion.isAsync).toBeFalsy();
+            expect(funktion.parameters).toHaveLength(0);
+            expect(funktion.body.toString()).toEqual('{ }');
         });
 
         it('should parse an async function declaration', () =>
         {
             const funktion = parser.parseFunction(FUNCTIONS.ASYNC_DECLARATION);
-
-            expect(funktion.name).toBe('name');
-            expect(funktion.isAsync).toBe(true);
-            expect(funktion.parameters.length).toBe(0);
-            expect(funktion.body).toBe('{ }');
+            expect(funktion.identifier).toEqual('name');
+            expect(funktion.isAsync).toBeTruthy();
+            expect(funktion.parameters).toHaveLength(0);
+            expect(funktion.body.toString()).toEqual('{ }');
         });
 
         it('should parse an expression function declaration', () =>
         {
-            const funktion = parser.parseFunction(FUNCTIONS.EXPRESSION);
+            const variable = parser.parseVariable(FUNCTIONS.EXPRESSION);
+            expect(variable.identifier).toEqual('name');
+            expect(variable.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(variable.type).toEqual('const');
+            expect(variable.initializer).toBeInstanceOf(ESFunction);
 
-            expect(funktion.name).toBe('name');
-            expect(funktion.isAsync).toBe(false);
-            expect(funktion.parameters.length).toBe(0);
-            expect(funktion.body).toBe('{ }');
+            const funktion = variable.initializer as ESFunction;
+            expect(funktion.identifier).toBeUndefined();
+            expect(funktion.isAsync).toBeFalsy();
+            expect(funktion.parameters).toHaveLength(0);
+            expect(funktion.body.toString()).toEqual('{ }');
         });
 
         it('should parse an async expression function declaration', () =>
         {
-            const funktion = parser.parseFunction(FUNCTIONS.ASYNC_EXPRESSION);
-
-            expect(funktion.name).toBe('name');
-            expect(funktion.isAsync).toBe(true);
-            expect(funktion.parameters.length).toBe(0);
-            expect(funktion.body).toBe('{ }');
+            const variable = parser.parseVariable(FUNCTIONS.ASYNC_EXPRESSION);
+            expect(variable.identifier).toEqual('name');
+            expect(variable.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(variable.type).toEqual('const');
+            expect(variable.initializer).toBeInstanceOf(ESFunction);
+            
+            const funktion = variable.initializer as ESFunction;
+            expect(funktion.identifier).toBeUndefined();
+            expect(funktion.isAsync).toBeTruthy();
+            expect(funktion.parameters).toHaveLength(0);
+            expect(funktion.body.toString()).toEqual('{ }');
         });
 
         it('should parse an arrow function declaration', () =>
         {
-            const funktion = parser.parseFunction(FUNCTIONS.ARROW);
-
-            expect(funktion.name).toBe('name');
-            expect(funktion.isAsync).toBe(false);
-            expect(funktion.parameters.length).toBe(0);
-            expect(funktion.body).toBe('{ }');
+            const variable = parser.parseVariable(FUNCTIONS.ARROW);
+            expect(variable.identifier).toEqual('name');
+            expect(variable.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(variable.type).toEqual('const');
+            expect(variable.initializer).toBeInstanceOf(ESArrowFunction);
+            
+            const funktion = variable.initializer as ESArrowFunction;
+            expect(funktion.identifier).toBeUndefined();
+            expect(funktion.isAsync).toBeFalsy();
+            expect(funktion.parameters).toHaveLength(0);
+            expect(funktion.body.toString()).toEqual('{ }');
         });
 
         it('should parse an arrow function expression declaration', () =>
         {
-            const funktion = parser.parseFunction(FUNCTIONS.ARROW_EXPRESSION);
-
-            expect(funktion.name).toBe('name');
-            expect(funktion.isAsync).toBe(false);
-            expect(funktion.parameters.length).toBe(0);
-            expect(funktion.body).toBe("'value'");
+            const variable = parser.parseVariable(FUNCTIONS.ARROW_EXPRESSION);
+            expect(variable.identifier).toEqual('name');
+            expect(variable.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(variable.type).toEqual('const');
+            expect(variable.initializer).toBeInstanceOf(ESArrowFunction);
+            
+            const funktion = variable.initializer as ESArrowFunction;
+            expect(funktion.identifier).toBeUndefined();
+            expect(funktion.isAsync).toBeFalsy();
+            expect(funktion.parameters).toHaveLength(0);
+            expect(funktion.body.toString()).toEqual("'value'");
         });
 
         it('should parse an arrow function argument declaration', () =>
         {
-            const funktion = parser.parseFunction(FUNCTIONS.ARROW_ARGUMENT);
-
-            expect(funktion.name).toBe('name');
-            expect(funktion.isAsync).toBe(false);
-            expect(funktion.parameters.length).toBe(1);
-            expect(funktion.body).toBe('arg');
+            const variable = parser.parseVariable(FUNCTIONS.ARROW_ARGUMENT);
+            expect(variable.identifier).toEqual('name');
+            expect(variable.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(variable.type).toEqual('const');
+            expect(variable.initializer).toBeInstanceOf(ESArrowFunction);
+            
+            const funktion = variable.initializer as ESArrowFunction;
+            expect(funktion.identifier).toBeUndefined();
+            expect(funktion.isAsync).toBeFalsy();
+            expect(funktion.parameters).toHaveLength(1);
+            expect(funktion.body.toString()).toEqual('arg');
         });
 
         it('should parse an async arrow function declaration', () =>
         {
-            const funktion = parser.parseFunction(FUNCTIONS.ASYNC_ARROW);
-
-            expect(funktion.name).toBe('name');
-            expect(funktion.isAsync).toBe(true);
-            expect(funktion.parameters.length).toBe(0);
-            expect(funktion.body).toBe('{ }');
+            const variable = parser.parseVariable(FUNCTIONS.ASYNC_ARROW);
+            expect(variable.identifier).toEqual('name');
+            expect(variable.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(variable.type).toEqual('const');
+            expect(variable.initializer).toBeInstanceOf(ESArrowFunction);
+            
+            const funktion = variable.initializer as ESArrowFunction;
+            expect(funktion.identifier).toBeUndefined();
+            expect(funktion.isAsync).toBeTruthy();
+            expect(funktion.parameters).toHaveLength(0);
+            expect(funktion.body.toString()).toEqual('{ }');
         });
 
         it('should parse a generator function declaration', () =>
         {
             const funktion = parser.parseFunction(FUNCTIONS.GENERATOR);
-
-            expect(funktion).toBeInstanceOf(ESGenerator);
-            expect(funktion.name).toBe('name');
-            expect(funktion.isAsync).toBe(false);
-            expect(funktion.parameters.length).toBe(0);
-            expect(funktion.body).toBe('{ }');
+            expect(funktion).toBeInstanceOf(ESGeneratorFunction);
+            expect(funktion.identifier).toEqual('name');
+            expect(funktion.isAsync).toBeFalsy();
+            expect(funktion.parameters).toHaveLength(0);
+            expect(funktion.body.toString()).toEqual('{ }');
         });
 
         it('should parse an async generator function declaration', () =>
         {
             const funktion = parser.parseFunction(FUNCTIONS.ASYNC_GENERATOR);
-
-            expect(funktion).toBeInstanceOf(ESGenerator);
-            expect(funktion.name).toBe('name');
-            expect(funktion.isAsync).toBe(true);
-            expect(funktion.parameters.length).toBe(0);
-            expect(funktion.body).toBe('{ }');
+            expect(funktion).toBeInstanceOf(ESGeneratorFunction);
+            expect(funktion.identifier).toEqual('name');
+            expect(funktion.isAsync).toBeTruthy();
+            expect(funktion.parameters).toHaveLength(0);
+            expect(funktion.body.toString()).toEqual('{ }');
         });
 
         it('should parse an expression generator function declaration', () =>
         {
-            const funktion = parser.parseFunction(FUNCTIONS.EXPRESSION_GENERATOR);
-
-            expect(funktion).toBeInstanceOf(ESGenerator);
-            expect(funktion.name).toBe('name');
-            expect(funktion.isAsync).toBe(false);
-            expect(funktion.parameters.length).toBe(0);
-            expect(funktion.body).toBe('{ }');
+            const variable = parser.parseVariable(FUNCTIONS.EXPRESSION_GENERATOR);
+            expect(variable.identifier).toEqual('name');
+            expect(variable.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(variable.type).toEqual('const');
+            expect(variable.initializer).toBeInstanceOf(ESGeneratorFunction);
+            
+            const funktion = variable.initializer as ESGeneratorFunction;
+            expect(funktion.identifier).toBeUndefined();
+            expect(funktion.isAsync).toBeFalsy();
+            expect(funktion.parameters).toHaveLength(0);
+            expect(funktion.body.toString()).toEqual('{ }');
         });
 
         it('should parse an async expression generator function declaration', () =>
         {
-            const funktion = parser.parseFunction(FUNCTIONS.ASYNC_EXPRESSION_GENERATOR);
-
-            expect(funktion).toBeInstanceOf(ESGenerator);
-            expect(funktion.name).toBe('name');
-            expect(funktion.isAsync).toBe(true);
-            expect(funktion.parameters.length).toBe(0);
-            expect(funktion.body).toBe('{ }');
+            const variable = parser.parseVariable(FUNCTIONS.ASYNC_EXPRESSION_GENERATOR);
+            expect(variable.identifier).toEqual('name');
+            expect(variable.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(variable.type).toEqual('const');
+            expect(variable.initializer).toBeInstanceOf(ESGeneratorFunction);
+            
+            const funktion = variable.initializer as ESGeneratorFunction;
+            expect(funktion.identifier).toBeUndefined();
+            expect(funktion.isAsync).toBeTruthy();
+            expect(funktion.parameters).toHaveLength(0);
+            expect(funktion.body.toString()).toEqual('{ }');
         });
 
         it('should parse a function with parameters', () =>
         {
             const funktion = parser.parseFunction(FUNCTIONS.PARAMETERS);
-            expect(funktion.name).toBe('name');
-            expect(funktion.isAsync).toBe(false);
-            expect(funktion.body).toBe('{ }');
+            expect(funktion.identifier).toEqual('name');
+            expect(funktion.isAsync).toBeFalsy();
+            expect(funktion.body.toString()).toEqual('{ }');
 
             const parameters = funktion.parameters;
-            expect(parameters.length).toBe(2);
+            expect(parameters).toHaveLength(2);
 
-            const first = parameters[0] as ESField;
-            expect(first).toBeInstanceOf(ESField);
-            expect(first.name).toBe('param1');
-            expect(first.value).toBe(undefined);
+            const first = parameters[0];
+            expect(first.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(first.binding.toString()).toEqual('param1');
+            expect(first.initializer).toBeUndefined();
 
-            const second = parameters[1] as ESField;
-            expect(second).toBeInstanceOf(ESField);
-            expect(second.name).toBe('param2');
-            expect(second.value).toBe(undefined);
+            const second = parameters[1];
+            expect(second.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(second.binding.toString()).toEqual('param2');
+            expect(second.initializer).toBeUndefined();
         });
 
         it('should parse a function with default parameters', () =>
         {
             const funktion = parser.parseFunction(FUNCTIONS.DEFAULT_PARAMETERS);
-            expect(funktion.name).toBe('name');
-            expect(funktion.isAsync).toBe(false);
-            expect(funktion.body).toBe('{ }');
+            expect(funktion.identifier).toEqual('name');
+            expect(funktion.isAsync).toBeFalsy();
+            expect(funktion.body.toString()).toEqual('{ }');
 
             const parameters = funktion.parameters;
-            expect(parameters.length).toBe(2);
+            expect(parameters).toHaveLength(2);
 
-            const first = parameters[0] as ESField;
-            expect(first).toBeInstanceOf(ESField);
-            expect(first.name).toBe('param1');
-            expect(first.value).toEqual(new ESExpression("'value1'"));
+            const first = parameters[0];
+            expect(first.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(first.binding.toString()).toEqual('param1');
+            expect(first.initializer).toBeInstanceOf(ESExpression);
+            expect(first.initializer?.toString(false)).toEqual("'value1'");
 
-            const second = parameters[1] as ESField;
-            expect(second).toBeInstanceOf(ESField);
-            expect(second.name).toBe('param2');
-            expect(second.value).toEqual(new ESExpression("true"));
+            const second = parameters[1];
+            expect(second.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(second.binding.toString()).toEqual('param2');
+            expect(second.initializer).toBeInstanceOf(ESExpression);
+            expect(second.initializer?.toString(false)).toEqual('true');
         });
 
         it('should parse a function with a rest parameter', () =>
         {
             const funktion = parser.parseFunction(FUNCTIONS.REST_PARAMETERS);
-            expect(funktion.name).toBe('name');
-            expect(funktion.isAsync).toBe(false);
-            expect(funktion.body).toBe('{ }');
+            expect(funktion.identifier).toEqual('name');
+            expect(funktion.isAsync).toBeFalsy();
+            expect(funktion.body.toString()).toEqual('{ }');
 
             const parameters = funktion.parameters;
-            expect(parameters.length).toBe(1);
+            expect(parameters).toHaveLength(1);
 
-            const first = parameters[0] as ESField;
-            expect(first).toBeInstanceOf(ESField);
-            expect(first.name).toBe('...param1');
-            expect(first.value).toEqual(undefined);
+            const first = parameters[0];
+            expect(first.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(first.binding.toString()).toEqual('...param1');
+            expect(first.initializer).toBeUndefined();
         });
 
         it('should parse a function with destructuring parameters', () =>
         {
             const funktion = parser.parseFunction(FUNCTIONS.DESTRUCTURING_PARAMETERS);
-            expect(funktion.name).toBe('name');
-            expect(funktion.isAsync).toBe(false);
-            expect(funktion.body).toBe('{ }');
+            expect(funktion.identifier).toEqual('name');
+            expect(funktion.isAsync).toBeFalsy();
+            expect(funktion.body.toString()).toEqual('{ }');
 
             const parameters = funktion.parameters;
-            expect(parameters.length).toBe(2);
+            expect(parameters).toHaveLength(2);
 
-            const firstParameter = parameters[0] as ESDestructuredObject;
-            expect(firstParameter).toBeInstanceOf(ESDestructuredObject);
-            expect(firstParameter.members.length).toBe(2);
+            const firstParameter = parameters[0];
+            expect(firstParameter.binding).toBeInstanceOf(ESObjectBinding);
+            expect(firstParameter.initializer).toBeUndefined();
 
-            const firstMember = firstParameter.members[0] as ESField;
-            expect(firstMember).toBeInstanceOf(ESField);
-            expect(firstMember.name).toBe('param1');
-            expect(firstMember.value).toBe(undefined);
+            const firstBinding = firstParameter.binding as ESObjectBinding;
+            expect(firstBinding.elements).toHaveLength(2);
+            expect(firstBinding.toString()).toEqual('{ param1, param2 }');
 
-            const secondMember = firstParameter.members[1] as ESField;
-            expect(secondMember).toBeInstanceOf(ESField);
-            expect(secondMember.name).toBe('param2');
-            expect(secondMember.value).toBe(undefined);
+            const firstMember = firstBinding.elements[0];
+            expect(firstMember.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(firstMember.binding.toString()).toEqual('param1');
+            expect(firstMember.initializer).toBeUndefined();
 
-            const secondParameter = parameters[1] as ESDestructuredArray;
-            expect(secondParameter).toBeInstanceOf(ESDestructuredArray);
-            expect(secondParameter.members.length).toBe(2);
+            const secondMember = firstBinding.elements[1];
+            expect(secondMember.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(secondMember.binding.toString()).toEqual('param2');
+            expect(secondMember.initializer).toBeUndefined();
 
-            const thirdMember = secondParameter.members[0] as ESField;
-            expect(thirdMember).toBeInstanceOf(ESField);
-            expect(thirdMember.name).toBe('param3');
-            expect(thirdMember.value).toBe(undefined);
+            const secondParameter = parameters[1];
+            expect(secondParameter.binding).toBeInstanceOf(ESArrayBinding);
+            expect(secondParameter.initializer).toBeUndefined();
 
-            const fourthMember = secondParameter.members[1] as ESField;
-            expect(fourthMember).toBeInstanceOf(ESField);
-            expect(fourthMember.name).toBe('param4');
-            expect(fourthMember.value).toBe(undefined);
+            const secondBinding = secondParameter.binding as ESObjectBinding;
+            expect(secondBinding.elements).toHaveLength(2);
+            expect(secondBinding.toString()).toEqual('[ param3, param4 ]');
+
+            const thirdMember = secondBinding.elements[0];
+            expect(thirdMember.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(thirdMember.binding.toString()).toEqual('param3');
+            expect(thirdMember.initializer).toBeUndefined();
+
+            const fourthMember = secondBinding.elements[1];
+            expect(fourthMember.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(fourthMember.binding.toString()).toEqual('param4');
+            expect(fourthMember.initializer).toBeUndefined();
         });
 
         it('should parse a function with destructuring default parameters', () =>
         {
             const funktion = parser.parseFunction(FUNCTIONS.DESTRUCTURING_DEFAULT_PARAMETERS);
-            expect(funktion.name).toBe('name');
-            expect(funktion.isAsync).toBe(false);
-            expect(funktion.body).toBe('{ }');
+            expect(funktion.identifier).toEqual('name');
+            expect(funktion.isAsync).toBeFalsy();
+            expect(funktion.body.toString()).toEqual('{ }');
 
             const parameters = funktion.parameters;
-            expect(parameters.length).toBe(2);
+            expect(parameters).toHaveLength(2);
 
-            const firstParameter = parameters[0] as ESDestructuredObject;
-            expect(firstParameter).toBeInstanceOf(ESDestructuredObject);
-            expect(firstParameter.members.length).toBe(2);
-            
-            const firstMember = firstParameter.members[0] as ESField;
-            expect(firstMember).toBeInstanceOf(ESField);
-            expect(firstMember.name).toBe('param1');
-            expect(firstMember.value).toBeInstanceOf(ESExpression);
+            const firstParameter = parameters[0];
+            expect(firstParameter.binding).toBeInstanceOf(ESObjectBinding);
+            expect(firstParameter.initializer).toBeUndefined();
 
-            const secondMember = firstParameter.members[1] as ESField;
-            expect(secondMember).toBeInstanceOf(ESField);
-            expect(secondMember.name).toBe('param2');
-            expect(secondMember.value).toBeInstanceOf(ESExpression);
+            const firstBinding = firstParameter.binding as ESObjectBinding;
+            expect(firstBinding.elements).toHaveLength(2);
+            expect(firstBinding.toString()).toEqual("{ param1 = 'value1', param2 = true }");
 
-            const secondParameter = parameters[1] as ESDestructuredArray;
-            expect(secondParameter).toBeInstanceOf(ESDestructuredArray);
-            expect(secondParameter.members.length).toBe(2);
-            
-            const thirdMember = secondParameter.members[0] as ESField;
-            expect(thirdMember).toBeInstanceOf(ESField);
-            expect(thirdMember.name).toBe('param3');
-            expect(thirdMember.value).toBeInstanceOf(ESExpression);
+            const firstMember = firstBinding.elements[0];
+            expect(firstMember.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(firstMember.binding.toString()).toEqual('param1');
+            expect(firstMember.initializer).toBeInstanceOf(ESExpression);
 
-            const fourthMember = secondParameter.members[1] as ESField;
-            expect(fourthMember).toBeInstanceOf(ESField);
-            expect(fourthMember.name).toBe('param4');
-            expect(fourthMember.value).toBeInstanceOf(ESExpression);
+            const secondMember = firstBinding.elements[1];
+            expect(secondMember.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(secondMember.binding.toString()).toEqual('param2');
+            expect(secondMember.initializer).toBeInstanceOf(ESExpression);
+
+            const secondParameter = parameters[1];
+            expect(secondParameter.binding).toBeInstanceOf(ESArrayBinding);
+            expect(secondParameter.initializer).toBeUndefined();
+
+            const secondBinding = secondParameter.binding as ESObjectBinding;
+            expect(secondBinding.elements).toHaveLength(2);
+            expect(secondBinding.toString()).toEqual("[ param3 = 'value3', param4 = true ]");
+
+            const thirdMember = secondBinding.elements[0];
+            expect(thirdMember.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(thirdMember.binding.toString()).toEqual('param3');
+            expect(thirdMember.initializer).toBeInstanceOf(ESExpression);
+
+            const fourthMember = secondBinding.elements[1];
+            expect(fourthMember.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(fourthMember.binding.toString()).toEqual('param4');
+            expect(fourthMember.initializer).toBeInstanceOf(ESExpression);
         });
 
         it('should parse a function with destructuring rest parameters', () =>
         {
             const funktion = parser.parseFunction(FUNCTIONS.DESTRUCTURING_REST_PARAMETERS);
-            expect(funktion.name).toBe('name');
-            expect(funktion.isAsync).toBe(false);
-            expect(funktion.body).toBe('{ }');
+            expect(funktion.identifier).toEqual('name');
+            expect(funktion.isAsync).toBeFalsy();
+            expect(funktion.body.toString()).toEqual('{ }');
 
             const parameters = funktion.parameters;
-            expect(parameters.length).toBe(2);
+            expect(parameters).toHaveLength(2);
 
-            const firstParameter = parameters[0] as ESDestructuredObject;
-            expect(firstParameter).toBeInstanceOf(ESDestructuredObject);
-            expect(firstParameter.members.length).toBe(2);
-            
-            const firstMember = firstParameter.members[0] as ESField;
-            expect(firstMember).toBeInstanceOf(ESField);
-            expect(firstMember.name).toBe('param1');
-            expect(firstMember.value).toBe(undefined);
+            const firstParameter = parameters[0];
+            expect(firstParameter.binding).toBeInstanceOf(ESObjectBinding);
+            expect(firstParameter.initializer).toBeUndefined();
 
-            const secondMember = firstParameter.members[1] as ESField;
-            expect(secondMember).toBeInstanceOf(ESField);
-            expect(secondMember.name).toBe('param2');
-            expect(secondMember.value).toBe(undefined);
+            const firstBinding = firstParameter.binding as ESObjectBinding;
+            expect(firstBinding.elements).toHaveLength(2);
+            expect(firstBinding.toString()).toEqual('{ param1, param2 }');
 
-            const secondParameter = parameters[1] as ESDestructuredArray;
-            expect(secondParameter).toBeInstanceOf(ESDestructuredArray);
-            expect(secondParameter.members.length).toBe(2);
-            
-            const thirdMember = secondParameter.members[0] as ESField;
-            expect(thirdMember).toBeInstanceOf(ESField);
-            expect(thirdMember.name).toBe('param3');
-            expect(thirdMember.value).toBe(undefined);
+            const firstMember = firstBinding.elements[0];
+            expect(firstMember.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(firstMember.binding.toString()).toEqual('param1');
+            expect(firstMember.initializer).toBeUndefined();
 
-            const fourthMember = secondParameter.members[1] as ESField;
-            expect(fourthMember).toBeInstanceOf(ESField);
-            expect(fourthMember.name).toBe('...param4');
-            expect(fourthMember.value).toBe(undefined);
+            const secondMember = firstBinding.elements[1];
+            expect(secondMember.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(secondMember.binding.toString()).toEqual('param2');
+            expect(secondMember.initializer).toBeUndefined();
+
+            const secondParameter = parameters[1];
+            expect(secondParameter.binding).toBeInstanceOf(ESArrayBinding);
+            expect(secondParameter.initializer).toBeUndefined();
+
+            const secondBinding = secondParameter.binding as ESObjectBinding;
+            expect(secondBinding.elements).toHaveLength(2);
+            expect(secondBinding.toString()).toEqual('[ param3, ...param4 ]');
+
+            const thirdMember = secondBinding.elements[0];
+            expect(thirdMember.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(thirdMember.binding.toString()).toEqual('param3');
+            expect(thirdMember.initializer).toBeUndefined();
+
+            const fourthMember = secondBinding.elements[1];
+            expect(fourthMember.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(fourthMember.binding.toString()).toEqual('...param4');
+            expect(fourthMember.initializer).toBeUndefined();
         });
 
         it('should parse a simple function body', () =>
         {
             const funktion = parser.parseFunction(FUNCTIONS.SIMPLE_BODY);
-            expect(funktion.name).toBe('name');
-            expect(funktion.isAsync).toBe(false);
-            expect(funktion.body).toBe("{ return 'value' ; }");
+
+            expect(funktion.identifier).toEqual('name');
+            expect(funktion.isAsync).toBeFalsy();
+            expect(funktion.body.toString()).toEqual("{ return 'value' ; }");
 
             const parameters = funktion.parameters;
-            expect(parameters.length).toBe(0);
+            expect(parameters).toHaveLength(0);
         });
 
         it('should parse a block function body', () =>
         {
             const funktion = parser.parseFunction(FUNCTIONS.BLOCK_BODY);
-            expect(funktion.name).toBe('name');
-            expect(funktion.isAsync).toBe(false);
-            expect(funktion.body).toBe("{ if ( true ) { return 'value' ; } }");
+
+            expect(funktion.identifier).toEqual('name');
+            expect(funktion.isAsync).toBeFalsy();
+            expect(funktion.body.toString()).toEqual("{ if ( true ) { return 'value' ; } }");
 
             const parameters = funktion.parameters;
-            expect(parameters.length).toBe(0);
+            expect(parameters).toHaveLength(0);
         });
 
         it('should parse function with a non reserved keyword as name', () =>
         {
             const funktion = parser.parseFunction(FUNCTIONS.KEYWORD_AS_NAME);
-            expect(funktion.name).toBe('as');
-            expect(funktion.isAsync).toBe(false);
-            expect(funktion.body).toBe("{ }");
+
+            expect(funktion.identifier).toEqual('as');
+            expect(funktion.isAsync).toBeFalsy();
+            expect(funktion.body.toString()).toEqual("{ }");
 
             const parameters = funktion.parameters;
-            expect(parameters.length).toBe(0);
+            expect(parameters).toHaveLength(0);
         });
     });
 
@@ -768,153 +854,194 @@ describe('parser/Parser', () =>
         it('should parse a simple class declaration', () =>
         {
             const clazz = parser.parseClass(CLASSES.DECLARATION);
-            expect(clazz.name).toBe('Name');
-            expect(clazz.parentName).toBe(undefined);
+            expect(clazz.identifier).toEqual('Name');
+            expect(clazz.parent).toBeUndefined();
 
             const members = clazz.members;
-            expect(members.length).toBe(0);
+            expect(members).toHaveLength(0);
         });
 
         it('should parse a simple class declaration with parent class', () =>
         {
             const clazz = parser.parseClass(CLASSES.EXTENDS);
-            expect(clazz.name).toBe('Name');
-            expect(clazz.parentName).toBe('Parent');
+            expect(clazz.identifier).toEqual('Name');
+            expect(clazz.parent).toEqual('Parent');
             
             const members = clazz.members;
-            expect(members.length).toBe(0);
+            expect(members).toHaveLength(0);
         });
 
         it('should parse an expression class declaration', () =>
         {
-            const clazz = parser.parseClass(CLASSES.EXPRESSION);
-            expect(clazz.name).toBe('name');
-            expect(clazz.parentName).toBe(undefined);
+            const variable = parser.parseVariable(CLASSES.EXPRESSION);
+            expect(variable.identifier).toEqual('name');
+            expect(variable.binding).toBeInstanceOf(ESIdentifierBinding);
+            expect(variable.type).toEqual('const');
+            expect(variable.initializer).toBeInstanceOf(ESClass);
+
+            const clazz = variable.initializer as ESClass;
+            expect(clazz.identifier).toBeUndefined();
+            expect(clazz.parent).toBeUndefined();
 
             const members = clazz.members;
-            expect(members.length).toBe(0);
+            expect(members).toHaveLength(0);
         });
 
         it('should parse class members', () =>
         {
             const clazz = parser.parseClass(CLASSES.MEMBERS);
-            expect(clazz.name).toBe('Name');
-            expect(clazz.parentName).toBe(undefined);
+            expect(clazz.identifier).toEqual('Name');
+            expect(clazz.parent).toBeUndefined();
 
             const members = clazz.members;
-            expect(members.length).toBe(21);
+            expect(members).toHaveLength(21);
 
-            const declarations = clazz.declarations;
-            expect(declarations.length).toBe(4);
+            const fields = clazz.fields;
+            expect(fields).toHaveLength(4);
 
-            expect(declarations[0].name).toBe('field1');
-            expect(declarations[0].isPrivate).toBe(true);
-            expect(declarations[0].isStatic).toBe(false);
-            expect(declarations[0].value?.definition).toBe("'value1'");
+            expect(fields[0].identifier).toEqual('field1');
+            expect(fields[0].visibility).toEqual('private');
+            expect(fields[0].location).toEqual('instance');
+            expect(fields[0].initializer?.toString(false)).toEqual("'value1'");
 
-            expect(declarations[1].name).toBe('field2');
-            expect(declarations[1].isPrivate).toBe(false);
-            expect(declarations[1].isStatic).toBe(false);
-            expect(declarations[1].value).toBe(undefined);
+            expect(fields[1].identifier).toEqual('field2');
+            expect(fields[1].visibility).toEqual('public');
+            expect(fields[1].location).toEqual('instance');
+            expect(fields[1].initializer).toBeUndefined();
 
-            expect(declarations[2].name).toBe('field3');
-            expect(declarations[2].isPrivate).toBe(true);
-            expect(declarations[2].isStatic).toBe(true);
-            expect(declarations[2].value?.definition).toBe('"value3"');
+            expect(fields[2].identifier).toEqual('field3');
+            expect(fields[2].visibility).toEqual('private');
+            expect(fields[2].location).toEqual('static');
+            expect(fields[2].initializer?.toString(false)).toEqual('"value3"');
 
-            expect(declarations[3].name).toBe('field4');
-            expect(declarations[3].isPrivate).toBe(false);
-            expect(declarations[3].isStatic).toBe(true);
-            expect(declarations[3].value).toBe(undefined);
+            expect(fields[3].identifier).toEqual('field4');
+            expect(fields[3].visibility).toEqual('public');
+            expect(fields[3].location).toEqual('static');
+            expect(fields[3].initializer).toBeUndefined();
 
             const getters = clazz.getters;
-            expect(getters.length).toBe(4);
+            expect(getters).toHaveLength(4);
 
-            expect(getters[0].name).toBe('getter1');
-            expect(getters[0].isPrivate).toBe(true);
-            expect(getters[0].isStatic).toBe(false);
+            expect(getters[0].identifier).toEqual('getter1');
+            expect(getters[0].visibility).toEqual('private');
+            expect(getters[0].location).toEqual('instance');
+            expect(getters[0].body.code).toEqual('{ return this.#field1 ; }');
 
-            expect(getters[1].name).toBe('getter2');
-            expect(getters[1].isPrivate).toBe(false);
-            expect(getters[1].isStatic).toBe(false);
+            expect(getters[1].identifier).toEqual('getter2');
+            expect(getters[1].visibility).toEqual('public');
+            expect(getters[1].location).toEqual('instance');
+            expect(getters[1].body.code).toEqual('{ return this.field2 ; }');
 
-            expect(getters[2].name).toBe('getter3');
-            expect(getters[2].isPrivate).toBe(true);
-            expect(getters[2].isStatic).toBe(true);
+            expect(getters[2].identifier).toEqual('getter3');
+            expect(getters[2].visibility).toEqual('private');
+            expect(getters[2].location).toEqual('static');
+            expect(getters[2].body.code).toEqual('{ return this.#field3 ; }');
 
-            expect(getters[3].name).toBe('getter4');
-            expect(getters[3].isPrivate).toBe(false);
-            expect(getters[3].isStatic).toBe(true);
+            expect(getters[3].identifier).toEqual('getter4');
+            expect(getters[3].visibility).toEqual('public');
+            expect(getters[3].location).toEqual('static');
+            expect(getters[3].body.code).toEqual('{ return this.field4 ; }');
 
             const setters = clazz.setters;
-            expect(setters.length).toBe(4);
+            expect(setters).toHaveLength(4);
 
-            expect(setters[0].name).toBe('setter1');
-            expect(setters[0].isPrivate).toBe(true);
-            expect(setters[0].isStatic).toBe(false);
+            expect(setters[0].identifier).toEqual('setter1');
+            expect(setters[0].visibility).toEqual('private');
+            expect(setters[0].location).toEqual('instance');
+            expect(setters[0].parameter.binding.toString()).toEqual('value');
+            expect(setters[0].body.code).toEqual('{ this.#field1 = value ; }');
 
-            expect(setters[1].name).toBe('setter2');
-            expect(setters[1].isPrivate).toBe(false);
-            expect(setters[1].isStatic).toBe(false);
+            expect(setters[1].identifier).toEqual('setter2');
+            expect(setters[1].visibility).toEqual('public');
+            expect(setters[1].location).toEqual('instance');
+            expect(setters[1].parameter.binding.toString()).toEqual('value');
+            expect(setters[1].body.code).toEqual('{ this.field2 = value ; }');
 
-            expect(setters[2].name).toBe('setter3');
-            expect(setters[2].isPrivate).toBe(true);
-            expect(setters[2].isStatic).toBe(true);
+            expect(setters[2].identifier).toEqual('setter3');
+            expect(setters[2].visibility).toEqual('private');
+            expect(setters[2].location).toEqual('static');
+            expect(setters[2].parameter.binding.toString()).toEqual('value');
+            expect(setters[2].body.code).toEqual('{ this.#field3 = value ; }');
 
-            expect(setters[3].name).toBe('setter4');
-            expect(setters[3].isPrivate).toBe(false);
-            expect(setters[3].isStatic).toBe(true);
+            expect(setters[3].identifier).toEqual('setter4');
+            expect(setters[3].visibility).toEqual('public');
+            expect(setters[3].location).toEqual('static');
+            expect(setters[3].parameter.binding.toString()).toEqual('value');
+            expect(setters[3].body.code).toEqual('{ this.field4 = value ; }');
 
-            const functions = clazz.functions;
-            expect(functions.length).toBe(6);
+            const construct = clazz.construct as ESConstructor;
+            expect(construct).toBeDefined();
+            expect(construct.identifier).toEqual('constructor');
+            expect(construct.visibility).toEqual('public');
+            expect(construct.location).toEqual('instance');
+            expect(construct.parameters).toHaveLength(2);
+            expect(construct.parameters[0].toString()).toEqual('field1');
+            expect(construct.parameters[1].toString()).toEqual('...field2');
+            expect(construct.body.toString()).toEqual('{ this.#field1 = field1 ; this.field2 = field2 ; }');
 
-            expect(functions[0].name).toBe('constructor');
-            expect(functions[0].isPrivate).toBe(false);
-            expect(functions[0].isStatic).toBe(false);
-            expect(functions[0].isAsync).toBe(false);
+            const methods = clazz.methods;
+            expect(methods).toHaveLength(8);
 
-            expect(functions[1].name).toBe('method1');
-            expect(functions[1].isPrivate).toBe(false);
-            expect(functions[1].isStatic).toBe(false);
-            expect(functions[1].isAsync).toBe(false);
+            expect(methods[0].identifier).toEqual('method1');
+            expect(methods[0].visibility).toEqual('public');
+            expect(methods[0].location).toEqual('instance');
+            expect(methods[0].isAsync).toBeFalsy();
+            expect(methods[0].parameters).toHaveLength(0);
+            expect(methods[0].body.toString()).toEqual('{ return this.#field1 ; }');
 
-            expect(functions[2].name).toBe('method2');
-            expect(functions[2].isPrivate).toBe(false);
-            expect(functions[2].isStatic).toBe(false);
-            expect(functions[2].isAsync).toBe(true);
+            expect(methods[1].identifier).toEqual('method2');
+            expect(methods[1].visibility).toEqual('public');
+            expect(methods[1].location).toEqual('instance');
+            expect(methods[1].isAsync).toBeTruthy();
+            expect(methods[1].parameters).toHaveLength(0);
+            expect(methods[1].body.toString()).toEqual('{ return this.field2 ; }');
 
-            expect(functions[3].name).toBe('method3');
-            expect(functions[3].isPrivate).toBe(false);
-            expect(functions[3].isStatic).toBe(true);
-            expect(functions[3].isAsync).toBe(false);
+            expect(methods[2].identifier).toEqual('method3');
+            expect(methods[2].visibility).toEqual('public');
+            expect(methods[2].location).toEqual('static');
+            expect(methods[2].isAsync).toBeFalsy();
+            expect(methods[2].parameters).toHaveLength(0);
+            expect(methods[2].body.toString()).toEqual('{ return this.#field3 ; }');
 
-            expect(functions[4].name).toBe('method4');
-            expect(functions[4].isPrivate).toBe(false);
-            expect(functions[4].isStatic).toBe(true);
-            expect(functions[4].isAsync).toBe(true);
+            expect(methods[3].identifier).toEqual('method4');
+            expect(methods[3].visibility).toEqual('public');
+            expect(methods[3].location).toEqual('static');
+            expect(methods[3].isAsync).toBeTruthy();
+            expect(methods[3].parameters).toHaveLength(0);
+            expect(methods[3].body.toString()).toEqual('{ return this.field4 ; }');
 
-            expect(functions[5].name).toBe('method5');
-            expect(functions[5].isPrivate).toBe(true);
-            expect(functions[5].isStatic).toBe(false);
-            expect(functions[5].isAsync).toBe(false);
+            expect(methods[4].identifier).toEqual('method5');
+            expect(methods[4].visibility).toEqual('private');
+            expect(methods[4].location).toEqual('instance');
+            expect(methods[4].isAsync).toBeFalsy();
+            expect(methods[4].parameters).toHaveLength(2);
+            expect(methods[4].parameters[0].toString()).toEqual('a');
+            expect(methods[4].parameters[1].toString()).toEqual('b');
+            expect(methods[4].body.toString()).toEqual('{ return a + b ; }');
 
-            const generators = clazz.generators;
-            expect(generators.length).toBe(3);
+            expect(methods[5]).toBeInstanceOf(ESGeneratorMethod);
+            expect(methods[5].identifier).toEqual('generator1');
+            expect(methods[5].visibility).toEqual('public');
+            expect(methods[5].location).toEqual('instance');
+            expect(methods[5].isAsync).toBeFalsy();
+            expect(methods[5].parameters).toHaveLength(0);
+            expect(methods[5].body.toString()).toEqual('{ yield 1 ; }');
 
-            expect(generators[0].name).toBe('generator1');
-            expect(generators[0].isPrivate).toBe(false);
-            expect(generators[0].isStatic).toBe(false);
-            expect(generators[0].isAsync).toBe(false);
+            expect(methods[6]).toBeInstanceOf(ESGeneratorMethod);
+            expect(methods[6].identifier).toEqual('generator2');
+            expect(methods[6].visibility).toEqual('public');
+            expect(methods[6].location).toEqual('instance');
+            expect(methods[6].isAsync).toBeTruthy();
+            expect(methods[6].parameters).toHaveLength(0);
+            expect(methods[6].body.toString()).toEqual('{ yield 1 ; }');
 
-            expect(generators[1].name).toBe('generator2');
-            expect(generators[1].isPrivate).toBe(false);
-            expect(generators[1].isStatic).toBe(false);
-            expect(generators[1].isAsync).toBe(true);
-
-            expect(generators[2].name).toBe('generator3');
-            expect(generators[2].isPrivate).toBe(false);
-            expect(generators[2].isStatic).toBe(true);
-            expect(generators[2].isAsync).toBe(true);
+            expect(methods[7]).toBeInstanceOf(ESGeneratorMethod);
+            expect(methods[7].identifier).toEqual('generator3');
+            expect(methods[7].visibility).toEqual('public');
+            expect(methods[7].location).toEqual('static');
+            expect(methods[7].isAsync).toBeTruthy();
+            expect(methods[7].parameters).toHaveLength(0);
+            expect(methods[7].body.toString()).toEqual('{ yield 1 ; }');
         });
     });
 
@@ -923,63 +1050,40 @@ describe('parser/Parser', () =>
         it('should parse an empty module', () =>
         {
             const module = parser.parse('');
-            
-            const members = module.members;
-            expect(members.length).toBe(0);
-        });
-
-        it('should parse omit root level expressions', () =>
-        {
-            const module = parser.parse(VALUES.EXPRESSION);
-            
-            const members = module.members;
-            expect(members.length).toBe(0);
+            expect(module.statements).toHaveLength(0);
         });
 
         it('should parse a module with terminated statements', () =>
         {
             const module = parser.parse(MODULES.TERMINATED);
-            
-            const members = module.members;
-            expect(members.length).toBe(14);
-
-            const imports = module.imports;
-            expect(imports.length).toBe(2);
-
-            const exports = module.exports;
-            expect(exports.length).toBe(4);
-
-            const declarations = module.declarations;
-            expect(declarations.length).toBe(4);
-
-            const functions = module.functions;
-            expect(functions.length).toBe(3);
-
-            const classes = module.classes;
-            expect(classes.length).toBe(1);
+            expect(module.statements).toHaveLength(16);
+            expect(module.imports).toHaveLength(2);
+            expect(module.exports).toHaveLength(4);
+            expect(module.expressions).toHaveLength(2);
+            expect(module.variables).toHaveLength(5);
+            expect(module.functions).toHaveLength(2);
+            expect(module.classes).toHaveLength(1);
         });
 
         it('should parse a module with unterminated statements', () =>
         {
             const module = parser.parse(MODULES.UNTERMINATED);
+            expect(module.statements).toHaveLength(10);
+            expect(module.imports).toHaveLength(2);
+            expect(module.exports).toHaveLength(3);
+            expect(module.expressions).toHaveLength(1);
+            expect(module.variables).toHaveLength(2);
+            expect(module.functions).toHaveLength(1);
+            expect(module.classes).toHaveLength(1);
+        });
+
+        it('should parse a module and regenerate its code', () =>
+        {
+            const module = parser.parse(MODULES.TERMINATED);
             
-            const members = module.members;
-            expect(members.length).toBe(9);
+            expect(module.toString()).toEqual(MODULES_STRINGS.TERMINATED);
 
-            const imports = module.imports;
-            expect(imports.length).toBe(2);
-
-            const exports = module.exports;
-            expect(exports.length).toBe(3);
-
-            const declarations = module.declarations;
-            expect(declarations.length).toBe(2);
-
-            const functions = module.functions;
-            expect(functions.length).toBe(1);
-
-            const classes = module.classes;
-            expect(classes.length).toBe(1);
+            // NOTE: Unterminated code is read only as it can not be outputted to working code
         });
     });
 });
